Move python modules into dedicated file
